PlotDesignMat: Error when no events near middle sample dataset, changed to plot around middle event instead
In timeexpanded case, plotDesignmat produced an error, if there was not experimental event within the plotting rate, i.e. if the vector "latix" was all zeros, then the setting of the ylims for plotting the designmatrix produced an error (i.e. because one of the ylims was an empty matrix). I changed the entire logic, so that now, the functions looks for a middle event in the data (that is part of model or not), and plots some designmatrix around that event. This makes sure that there is "something" to see in the visualized Designmatrix and avoids the error above.