Changed "zeros ignored" warning in uf_predictContinuous
When computing the values at which to evaluate continuous linear predictors, we offer the average or auto-spacing at percentiles as simple options. We have the problem that we cannot differentiate "real" zero-values of the predictor (i.e. rows in which an actual event occured that might be really zero) from other zeros in the design matrix. This will bias estimations of the mean or perceptiles. The previous warning message was cryptic and confused users. Rephrased the warning.