Critical bugfix to included ERPLAB subfunction: uf_continuousArtifactDetect
Critical error in an apparently outdated (and at that time "crappy") ERPLAB subfunction basicrap.m means that one artifacts were detected in more than one channel, the outputs were wrong, with artifacts being not correctly detected. Detection results were only correct if only one channel was specified. This is a quickfix to the old version of the basicrap function we are using, but we need to auto-include current ERPLAB versions.